move shared components
