+++ src/components/AcceleratorReports/RejectedReportMessage.tsx
